HomeViewController tableview initialize test passing
