+++ Lodjinha/Lodjinha/HomeViewController.xib
@@ -1,0 +1,38 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<document type="com.apple.InterfaceBuilder3.CocoaTouch.XIB" version="3.0" toolsVersion="14460.31" targetRuntime="iOS.CocoaTouch" propertyAccessControl="none" useAutolayout="YES" useTraitCollections="YES" colorMatched="YES">
+    <device id="retina4_7" orientation="portrait">
+        <adaptation id="fullscreen"/>
+    </device>
+    <dependencies>
+        <deployment identifier="iOS"/>
+        <plugIn identifier="com.apple.InterfaceBuilder.IBCocoaTouchPlugin" version="14460.20"/>
+        <capability name="documents saved in the Xcode 8 format" minToolsVersion="8.0"/>
+    </dependencies>
+    <objects>
+        <placeholder placeholderIdentifier="IBFilesOwner" id="-1" userLabel="File's Owner" customClass="HomeViewController" customModule="Lodjinha" customModuleProvider="target">
+            <connections>
+                <outlet property="tableView" destination="LcX-aX-Sys" id="Jew-ZQ-rom"/>
+                <outlet property="view" destination="iN0-l3-epB" id="7Sa-5r-o7e"/>
+            </connections>
+        </placeholder>
+        <placeholder placeholderIdentifier="IBFirstResponder" id="-2" customClass="UIResponder"/>
+        <view contentMode="scaleToFill" id="iN0-l3-epB">
+            <rect key="frame" x="0.0" y="0.0" width="375" height="667"/>
+            <autoresizingMask key="autoresizingMask" widthSizable="YES" heightSizable="YES"/>
+            <subviews>
+                <tableView clipsSubviews="YES" contentMode="scaleToFill" alwaysBounceVertical="YES" style="plain" separatorStyle="default" rowHeight="-1" estimatedRowHeight="-1" sectionHeaderHeight="28" sectionFooterHeight="28" translatesAutoresizingMaskIntoConstraints="NO" id="LcX-aX-Sys">
+                    <rect key="frame" x="0.0" y="0.0" width="375" height="667"/>
+                    <color key="backgroundColor" white="1" alpha="1" colorSpace="custom" customColorSpace="genericGamma22GrayColorSpace"/>
+                </tableView>
+            </subviews>
+            <color key="backgroundColor" red="1" green="1" blue="1" alpha="1" colorSpace="custom" customColorSpace="sRGB"/>
+            <constraints>
+                <constraint firstItem="LcX-aX-Sys" firstAttribute="bottom" secondItem="iN0-l3-epB" secondAttribute="bottom" id="4vQ-VJ-BZl"/>
+                <constraint firstItem="LcX-aX-Sys" firstAttribute="trailing" secondItem="iN0-l3-epB" secondAttribute="trailing" id="Wrz-wr-MgJ"/>
+                <constraint firstItem="LcX-aX-Sys" firstAttribute="leading" secondItem="iN0-l3-epB" secondAttribute="leading" id="l0O-9b-iRV"/>
+                <constraint firstItem="LcX-aX-Sys" firstAttribute="top" secondItem="iN0-l3-epB" secondAttribute="top" id="zI0-Ze-qIf"/>
+            </constraints>
+            <point key="canvasLocation" x="53.600000000000001" y="48.125937031484263"/>
+        </view>
+    </objects>
+</document>
