--- Lodjinha/Lodjinha/ViewController.swift
@@ -1,20 +1,0 @@
-//
-//  ViewController.swift
-//  Lodjinha
-//
-//  Created by Douglas da Silva Santos on 24/03/19.
-//  Copyright © 2019 Douglas da Silva Santos. All rights reserved.
-//
-
-import UIKit
-
-class ViewController: UIViewController {
-
-    override func viewDidLoad() {
-        super.viewDidLoad()
-        // Do any additional setup after loading the view, typically from a nib.
-    }
-
-
-}
-
